buildatlas/apps/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
